pyproject.toml
[tool.poetry]
name = "mltoolsbot"
version = "1.0.1"
description = "Repository for telegram bot which helps user with machine learning through communication and commands"
authors = ["Alexander <as.frantsev@gmail.com>"]
license = "MIT"
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.2"
python-telegram-bot = "^20.7"
aiolimiter = "^1.1.0"
python-dotenv = "^1.0.0"
redis = "^5.2.1"
anthropic = "^0.42.0"
elevenlabs = "^1.50.6"
annotated-types = "^0.7.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^24.2.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
mltoolsbot = "mltoolsbot.run:main"


[tool.black]
line-length = 88
target-version = ['py311']
extend-exclude = '''
(
  migrations   # Django DB migrations
)
'''


[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
skip_gitignore = true
skip_glob = ["**/migrations/*", "**/settings/*"]
src_paths = ["<your_code_dir>"]